// angular-src/src/app/models/incentiveentry.ts
export class IncentiveEntry {
    CustomerVisit: boolean;
    ContractResign: boolean;
    AddRate: boolean;
    LteUpgrade: boolean;
    LandlineToCellConversion: boolean;
    SystemReprogram: boolean;
    ServicePerformed: boolean;
    SitePickup: boolean;
    NewSite: boolean;
    SystemTransfer: boolean;
    Other: boolean;
}